+++ scripts/actions.js
@@ -7,7 +7,7 @@ import {
     waterExtraActionRestriction,
 } from "./action-layout.js";
 import { getWeapons } from "./data.js";
-import { installHudButtonIcon } from "./hud-buttons.js";
+import { installDelayedHudTooltip, installHudButtonIcon } from "./hud-buttons.js";
 import { buildNinjoGiri } from "./identity.js";
 import { getPersuadeOptions, openDicePicker, openGenericRoll, openPersuadeRoll, openWeaponStrike } from "./rolls.js";
 import {
@@ -496,6 +496,10 @@ export function createActionPanels(ARGON, { L5R5eEquipmentPanelButton }, { L5R5e
                     },
                 ],
             };
+        }
+
+        async activateTooltipListeners() {
+            installDelayedHudTooltip(this);
         }
 
         async _onLeftClick() {